Refactoring: Move all tick logic into TickUtils
